# Conflicts:
#	embassy-boot/boot/Cargo.toml
#	embassy-boot/nrf/Cargo.toml
#	embassy-embedded-hal/Cargo.toml
#	embassy-nrf/Cargo.toml
#	embassy-nrf/src/lib.rs
#	embassy-stm32/src/usart/mod.rs

+++ embassy-rp/src/rtc/mod.rs
@@ -26,12 +26,7 @@ impl<'d, T: Instance> RealTimeClock<'d, T> {
         into_ref!(inner);
 
         // Set the RTC divider
-        unsafe {
-            inner
-                .regs()
-                .clkdiv_m1()
-                .write(|w| w.set_clkdiv_m1(clk_rtc_freq() as u16 - 1))
-        };
+        unsafe { inner.regs().clkdiv_m1().write(|w| w.set_clkdiv_m1(clk_rtc_freq() - 1)) };
 
         let mut result = Self { inner };
         result.set_leap_year_check(true); // should be on by default, make sure this is the case.
@@ -126,7 +121,7 @@ impl<'d, T: Instance> RealTimeClock<'d, T> {
     /// # #[cfg(not(feature = "chrono"))]
     /// # fn main() {
     /// # use embassy_rp::rtc::{RealTimeClock, DateTimeFilter};
-    /// # let mut real_time_clock: RealTimeClock = unsafe { core::mem::zeroed() };
+    /// # let mut real_time_clock: RealTimeClock<embassy_rp::peripherals::RTC> = unsafe { core::mem::zeroed() };
     /// let now = real_time_clock.now().unwrap();
     /// real_time_clock.schedule_alarm(
     ///     DateTimeFilter::default()
